# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CollectionAttempt": {
        "additionalProperties": false,
        "description": "Sanitized telemetry for one provider or fallback-layer attempt.",
        "properties": {
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "number"
          },
          "error_code": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "http_status": {
            "anyOf": [
              {
                "maximum": 599.0,
                "minimum": 100.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Http Status"
          },
          "item_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Item Count",
            "type": "integer"
          },
          "layer": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Layer",
            "type": "string"
          },
          "operation": {
            "enum": [
              "authenticate",
              "discover",
              "listing",
              "item"
            ],
            "title": "Operation",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "succeeded",
              "failed",
              "stopped"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retryable"
          },
          "sequence": {
            "minimum": 0.0,
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "sequence",
          "layer",
          "operation",
          "outcome",
          "duration_ms"
        ],
        "title": "CollectionAttempt",
        "type": "object"
      },
      "CollectionErrorEvidence": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "message": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "CollectionErrorEvidence",
        "type": "object"
      },
      "CollectorReadiness": {
        "additionalProperties": false,
        "properties": {
          "authorized": {
            "title": "Authorized",
            "type": "boolean"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "source": {
            "enum": [
              "dazhong",
              "reddit",
              "guardian",
              "pikabu"
            ],
            "title": "Source",
            "type": "string"
          },
          "state": {
            "enum": [
              "disabled",
              "unconfigured",
              "unauthorized",
              "ready"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "source",
          "state",
          "enabled",
          "configured",
          "authorized"
        ],
        "title": "CollectorReadiness",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "story_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Id"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "trace_id"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "SourceItemIngestionOutcome": {
        "enum": [
          "ingested",
          "duplicate",
          "failed"
        ],
        "title": "SourceItemIngestionOutcome",
        "type": "string"
      },
      "SourceItemIngestionResult": {
        "additionalProperties": false,
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "external_id": {
            "minLength": 1,
            "title": "External Id",
            "type": "string"
          },
          "outcome": {
            "$ref": "#/components/schemas/SourceItemIngestionOutcome"
          },
          "story_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Id"
          }
        },
        "required": [
          "external_id",
          "outcome"
        ],
        "title": "SourceItemIngestionResult",
        "type": "object"
      },
...
      "SourceRun": {
        "additionalProperties": false,
        "properties": {
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/CollectionAttempt"
            },
            "maxItems": 1000,
            "title": "Attempts",
            "type": "array"
          },
          "collection_errors": {
            "items": {
              "$ref": "#/components/schemas/CollectionErrorEvidence"
            },
            "maxItems": 100,
            "title": "Collection Errors",
            "type": "array"
          },
          "collector_outcome": {
            "anyOf": [
              {
                "enum": [
                  "disabled",
                  "unconfigured",
                  "unauthorized",
                  "succeeded",
                  "partial",
                  "failed",
                  "stopped_captcha"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collector Outcome"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duplicate_count": {
            "readOnly": true,
            "title": "Duplicate Count",
            "type": "integer"
          },
          "failed_count": {
            "readOnly": true,
            "title": "Failed Count",
            "type": "integer"
          },
          "failure_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "readOnly": true,
            "title": "Failure Rate"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "ingested_count": {
            "readOnly": true,
            "title": "Ingested Count",
            "type": "integer"
          },
          "item_results": {
            "items": {
              "$ref": "#/components/schemas/SourceItemIngestionResult"
            },
            "maxItems": 100,
            "title": "Item Results",
            "type": "array"
          },
          "items_discovered": {
            "default": 0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Items Discovered",
            "type": "integer"
          },
          "request": {
            "$ref": "#/components/schemas/SourceRunRequest"
          },
          "run_error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CollectionErrorEvidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/SourceRunStatus",
            "default": "queued"
          },
          "trace_id": {
            "format": "uuid",
            "title": "Trace Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "request",
          "ingested_count",
          "duplicate_count",
          "failed_count",
          "failure_rate"
        ],
        "title": "SourceRun",
        "type": "object"
      },
      "SourceRunReadiness": {
        "additionalProperties": false,
        "properties": {
          "collectors": {
            "items": {
              "$ref": "#/components/schemas/CollectorReadiness"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Collectors",
            "type": "array"
          }
        },
        "required": [
          "collectors"
        ],
        "title": "SourceRunReadiness",
        "type": "object"
      },
      "SourceRunRequest": {
        "additionalProperties": false,
        "properties": {
          "emotion": {
            "default": "neutral",
            "minLength": 1,
            "title": "Emotion",
            "type": "string"
          },
          "limit": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "pitch": {
            "default": 0.0,
            "maximum": 12.0,
            "minimum": -12.0,
            "title": "Pitch",
            "type": "number"
          },
          "requested_by": {
            "minLength": 1,
            "title": "Requested By",
            "type": "string"
          },
          "source": {
            "enum": [
              "dazhong",
              "reddit",
              "guardian",
              "pikabu"
            ],
            "title": "Source",
            "type": "string"
          },
          "speaker_id": {
            "default": "narrator",
            "minLength": 1,
            "title": "Speaker Id",
            "type": "string"
          },
          "speed": {
            "default": 1.0,
            "maximum": 1.65,
            "minimum": 0.6,
            "title": "Speed",
            "type": "number"
          },
          "style": {
            "default": "clear, accurate short-video narration",
            "minLength": 1,
            "title": "Style",
            "type": "string"
          },
          "target_duration_seconds": {
            "default": 90,
            "maximum": 600.0,
            "minimum": 15.0,
            "title": "Target Duration Seconds",
            "type": "integer"
          },
          "target_language": {
            "default": "zh-CN",
            "minLength": 1,
            "title": "Target Language",
            "type": "string"
          }
        },
        "required": [
          "source",
          "requested_by"
        ],
        "title": "SourceRunRequest",
        "type": "object"
      },
      "SourceRunStatus": {
        "enum": [
          "queued",
          "collecting",
          "ingesting",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "SourceRunStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Review-gated multilingual content-to-audio pipeline.",
    "title": "god-news",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/source-runs": {
      "get": {
        "operationId": "listSourceRuns",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "dazhong",
                    "reddit",
                    "guardian",
                    "pikabu"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "run_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SourceRunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceRun"
                  },
                  "title": "Response Listsourceruns",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Source Runs",
        "tags": [
          "source-runs"
        ]
      },
      "post": {
        "operationId": "startSourceRun",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start Source Run",
        "tags": [
          "source-runs"
        ]
      }
    },
    "/api/v1/source-runs/{run_id}": {
      "get": {
        "operationId": "getSourceRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Source Run",
        "tags": [
          "source-runs"
        ]
      }
    },
    "/api/v1/source-runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancelSourceRun",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Cancel Source Run",
        "tags": [
          "source-runs"
        ]
      }
    },
    "/api/v1/sources/collectors": {
      "get": {
        "operationId": "getSourceCollectorReadiness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceRunReadiness"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Collector Readiness",
        "tags": [
          "source-runs"
        ]
      }
    },
...
  }
}